+++ src/main/java/com/api/cossc/service/quiz/QuizServiceImpl.java
@@ -1,4 +1,4 @@
-package com.api.cossc.service;
+package com.api.cossc.service.quiz;
 
 import com.api.cossc.domain.DailyQuizEntity;
 import com.api.cossc.domain.QuizEntity;
